Add Node.remove
This lets long running processes dump data (via `toJSONSubgraph`) and then
remove refs to avoid leaking memory.

Also
- cleanup (split up heimdalljs into constituent parts)
